% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_kdata.R
\name{as_kdata}
\alias{as_kdata}
\alias{as_kdata.data.frame}
\alias{as_kdata.sf}
\alias{as_kdata.sfc}
\alias{as_kdata.matrix}
\alias{as_kdata.list}
\alias{as_kdata.default}
\title{Convert an input object to a data frame (or sf) for kleaflet}
\usage{
as_kdata(x, ...)

\method{as_kdata}{data.frame}(x, ...)

\method{as_kdata}{sf}(x, ...)

\method{as_kdata}{sfc}(x, ...)

\method{as_kdata}{matrix}(x, ...)

\method{as_kdata}{list}(x, ...)

\method{as_kdata}{default}(x, ...)
}
\arguments{
\item{x}{Object to convert.}

\item{...}{Unused, for method extensions.}
}
\value{
A data frame or an \code{sf} object.
}
\description{
\code{as_kdata()} is the S3 entry point that turns whatever you hand to
\code{\link[=kleaflet]{kleaflet()}} into a data frame, or an \code{sf} object in longitude/latitude
(WGS84). Add a method to teach kleaflet a new input class. A method may
set the attributes \code{kleaf_lon}, \code{kleaf_lat} (default coordinate columns)
and \code{kleaf_type} (default layer type) on the result; they are only used
when the call does not specify \code{lon}, \code{lat} or \code{type}.
}
\details{
Built-in methods: \code{data.frame} (and tibbles), \code{sf} (reprojected to WGS84),
\code{sfc}, two-column \code{matrix} (longitude, latitude), \code{list} of equal-length
vectors.
}
\examples{
as_kdata(matrix(c(2.35, 48.85, -0.12, 51.5), ncol = 2, byrow = TRUE))
as_kdata(list(lon = c(2.35, -0.12), lat = c(48.85, 51.5)))

}
