% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{kleaf_save}
\alias{kleaf_save}
\title{Save a kleaflet to an HTML file}
\usage{
kleaf_save(p, file, selfcontained = TRUE, ...)
}
\arguments{
\item{p}{A \code{kleaflet} or \code{leaflet} map.}

\item{file}{File to save to (\code{.html}).}

\item{selfcontained}{Whether to embed everything in the file (needs
pandoc).}

\item{...}{Passed to \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}}.}
}
\value{
The file name, invisibly.
}
\description{
Thin wrapper around \code{\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget()}} that accepts a
\code{kleaflet}.
}
