% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kleaflet.r-package.R
\docType{package}
\name{kleaflet.r-package}
\alias{kleaflet.r}
\alias{kleaflet.r-package}
\title{kleaflet.r: Concise, Composable Interactive Maps from a Single Call}
\description{
Draw a complete interactive 'leaflet' map (coordinates, colour palettes, legends, popups, labels, groups, base map, view) with a single short call, and compose maps with the '+' operator through the 'kleaflet' S3 class. Layer types, themes and palettes live in extensible registries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kbosirany/kleaflet.r}
  \item \url{https://kbosirany.github.io/kleaflet.r/}
  \item Report bugs at \url{https://github.com/kbosirany/kleaflet.r/issues}
}

}
\author{
\strong{Maintainer}: Kevin Bosirany Orlando \email{kevinbosirany@gmail.com} (\href{https://orcid.org/0009-0009-2784-3108}{ORCID})

}
\keyword{internal}
